docs: rfc: 0099: multi column log and multi leader
A multi-leader setup offers the advantage that writes on each leader do
not block one another. However, the drawback is that there is no strict
guarantee of linearizability between different columns (leaders).

Therefore, this configuration is useful in scenarios where transactions
are not strictly dependent on each other. Such as a cross-data-center
deployment.